Add upload status to app-level panel
